perf(sched): Optimize rr_per_prio RAM usage from size_t to uint8_t

+++ src/syntropic/sched/syn_sched.c
@@ -203,7 +203,7 @@ bool syn_sched_run(SYN_Sched *sched)
          * state to DEFERRED before yielding — no RR advance). */
         if (best_task->state != (uint8_t)SYN_TASK_DEFERRED) {
             const size_t next = executed_idx + 1;
-            sched->rr_per_prio[best_prio] = (next >= n) ? 0 : next;
+            sched->rr_per_prio[best_prio] = (next >= n) ? 0 : (uint8_t)next;
         }
         break; /* One useful execution per tick */
     }
